[webapp] move profile parsing to feature
